requirements.txt
instaloader>=4.14
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
pyyaml>=6.0
python-dotenv>=1.0.0
